+++ blakcomp/codegen.h
@@ -45,8 +45,8 @@ void OutputBaseExpression(int outfile, expr_type expr);
 void BackpatchGoto(int outfile, int source, int destination);
 
 void codegen_error(const char *fmt, ...);
-int const_to_int(const_type c);
-int set_source_id(opcode_type *opcode, int sourcenum, expr_type e);
+__int64 const_to_int(const_type c);
+__int64 set_source_id(opcode_type *opcode, int sourcenum, expr_type e);
 int set_dest_id(opcode_type *opcode, id_type id);
 int is_base_level(expr_type e);
 id_type make_temp_var(int idnum);
